+++ src/Controls/src/Core/Button/Button.cs
@@ -459,7 +459,7 @@ namespace Microsoft.Maui.Controls
 		/// <param name="textTransform">The transform to apply to <paramref name="source"/>.</param>
 		/// <returns>The transformed text.</returns>
 		public virtual string UpdateFormsText(string source, TextTransform textTransform)
-			=> TextTransformUtilites.GetTransformedText(source, textTransform);
+			=> TextTransformUtilities.GetTransformedText(source, textTransform);
 
 		void ICommandElement.CanExecuteChanged(object sender, EventArgs e) =>
 			RefreshIsEnabledProperty();
