+++ src/Controls/src/Core/Platform/Windows/Extensions/FormattedStringExtensions.cs
@@ -123,7 +123,7 @@ namespace Microsoft.Maui.Controls.Platform
 
 			var transform = span.TextTransform != TextTransform.Default ? span.TextTransform : defaultTextTransform;
 
-			var text = TextTransformUtilites.GetTransformedText(span.Text, transform);
+			var text = TextTransformUtilities.GetTransformedText(span.Text, transform);
 
 			var run = new Run { Text = text ?? string.Empty };
