+++ src/Controls/src/Core/Platform/iOS/Extensions/ButtonExtensions.cs
@@ -84,7 +84,7 @@ namespace Microsoft.Maui.Controls.Platform
 
 		public static void UpdateText(this UIButton platformButton, Button button)
 		{
-			var text = TextTransformUtilites.GetTransformedText(button.Text, button.TextTransform);
+			var text = TextTransformUtilities.GetTransformedText(button.Text, button.TextTransform);
 			platformButton.SetTitle(text, UIControlState.Normal);
 
 			// The TitleLabel retains its previous text value even after a new value is assigned. As a result, the label does not display the updated text and reverts to the old value when the button is re-measured
