+++ src/Controls/src/Core/PublicAPI/net-tizen/PublicAPI.Shipped.txt
@@ -2366,8 +2366,6 @@
 ~static Microsoft.Maui.Controls.Internals.Registrar.RegisterRenderers(Microsoft.Maui.Controls.HandlerAttribute[] attributes) -> void
 ~static Microsoft.Maui.Controls.Internals.Registrar.RegisterRendererToHandlerShim(System.Func<object, Microsoft.Maui.IViewHandler> handlerShim) -> void
 ~static Microsoft.Maui.Controls.Internals.ResourceLoader.ResourceProvider2.get -> System.Func<Microsoft.Maui.Controls.Internals.ResourceLoader.ResourceLoadingQuery, Microsoft.Maui.Controls.Internals.ResourceLoader.ResourceLoadingResponse>
-~static Microsoft.Maui.Controls.Internals.TextTransformUtilites.GetTransformedText(string source, Microsoft.Maui.TextTransform textTransform) -> string
-~static Microsoft.Maui.Controls.Internals.TextTransformUtilites.SetPlainText(Microsoft.Maui.Controls.InputView inputView, string platformText) -> void
 ~static Microsoft.Maui.Controls.KnownColor.Accent.get -> Microsoft.Maui.Graphics.Color
 ~static Microsoft.Maui.Controls.KnownColor.Default.get -> Microsoft.Maui.Graphics.Color
 ~static Microsoft.Maui.Controls.KnownColor.SetAccent(Microsoft.Maui.Graphics.Color value) -> void
